requirements.txt
Django>=3.0,<4.1.8
psycopg2-binary>=2.8
djangorestframework==3.12.2
Pillow==9.5.0
requests==2.25.0
djangorestframework-simplejwt==4.6.0
PyJWT==1.7.1
django-request==1.5.6
urllib3==1.26.2
jsonify==0.5
python-dateutil==2.8.1
django-seed==0.2.2
python-dotenv==0.15.0
mysqlclient==2.0.2
pytz==2023.3
django-filter == 23.2
instapy

